# app/routes/api_exercises.py
import sqlite3
from typing import Annotated, Any

from fastapi import APIRouter, HTTPException, Query

from app.repositories.exercises import (
    ActiveExerciseWeightError,
    create_exercise,
    get_exercise,
    list_exercises,
    normalize_exercise_name,
    reorder_exercises,
    replace_exercise_weights,
    update_exercise,
)
from app.services.analysis_service import list_exercise_profiles
from app.services.stats_service import build_exercise_stats, parse_limit
from app.schemas import (
    ExerciseCreateRequest,
    ExerciseMutationResponse,
    ExerciseOrderUpdateRequest,
    ExerciseProfilesResponse,
    ExerciseStatsResponseModel,
    ExerciseUpdateRequest,
    ExerciseWeightsResponse,
    ExerciseWeightsUpdateRequest,
    ExercisesResponse,
)


router = APIRouter(prefix="/api/v1/exercises", tags=["exercises"])
profiles_router = APIRouter(prefix="/api/v1/exercise-profiles", tags=["exercises"])


def clean_exercise_name(name: str) -> str:
    return normalize_exercise_name(name)


@router.get("", response_model=ExercisesResponse)
def get_exercises(
    include_inactive: bool = False,
) -> dict[str, list[dict[str, Any]]]:
    return {
        "exercises": list_exercises(include_inactive=include_inactive),
    }


@router.get("/{exercise_id}/stats", response_model=ExerciseStatsResponseModel)
def get_exercise_stats_endpoint(
    exercise_id: int,
    limit: Annotated[str | None, Query()] = "30",
) -> dict[str, Any]:
    parsed_limit = parse_limit(limit, default=30)
    stats = build_exercise_stats(exercise_id, limit=parsed_limit)

    if stats is None:
        raise HTTPException(status_code=404, detail="Exercise not found.")

    return stats


@router.post("", response_model=ExerciseMutationResponse)
def create_exercise_endpoint(
    payload: ExerciseCreateRequest,
) -> dict[str, Any]:
    name = clean_exercise_name(payload.name)
    if not name:
        raise HTTPException(status_code=400, detail="Exercise name is required.")

    try:
        exercise, created = create_exercise(
            name,
            is_active=payload.is_active,
            profile_key=payload.profile_key,
            weights=payload.weights,
        )
    except ActiveExerciseWeightError as exc:
        raise HTTPException(status_code=409, detail=str(exc)) from exc
    except ValueError as exc:
        raise HTTPException(status_code=400, detail=str(exc)) from exc
    except sqlite3.IntegrityError as exc:
        raise HTTPException(
            status_code=409,
            detail="Exercise name already exists.",
        ) from exc

    return {
        "exercise": exercise,
        "created": created,
    }


@router.patch("/{exercise_id}", response_model=ExerciseMutationResponse)
def update_exercise_endpoint(
    exercise_id: int,
    payload: ExerciseUpdateRequest,
) -> dict[str, Any]:
    changed_fields = payload.model_fields_set
    if not changed_fields:
        raise HTTPException(status_code=400, detail="No exercise fields provided.")

    name = None
    if "name" in changed_fields:
        name = clean_exercise_name(payload.name or "")
        if not name:
            raise HTTPException(status_code=400, detail="Exercise name is required.")

    if get_exercise(exercise_id) is None:
        raise HTTPException(status_code=404, detail="Exercise not found.")

    try:
        exercise = update_exercise(
            exercise_id,
            name=name,
            is_active=payload.is_active if "is_active" in changed_fields else None,
            profile_key=(
                payload.profile_key
                if "profile_key" in changed_fields
                else None
            ),
        )
    except ActiveExerciseWeightError as exc:
        raise HTTPException(status_code=409, detail=str(exc)) from exc
    except ValueError as exc:
        raise HTTPException(status_code=400, detail=str(exc)) from exc
    except sqlite3.IntegrityError as exc:
        raise HTTPException(
            status_code=409,
            detail="Exercise name already exists.",
        ) from exc

    if exercise is None:
        raise HTTPException(status_code=404, detail="Exercise not found.")

    return {"exercise": exercise}


@router.put("/order", response_model=ExercisesResponse)
def reorder_exercises_endpoint(
    payload: ExerciseOrderUpdateRequest,
) -> dict[str, Any]:
    try:
        exercises = reorder_exercises(payload.exercise_ids)
    except ValueError as exc:
        raise HTTPException(status_code=400, detail=str(exc)) from exc

    return {"exercises": exercises}


@router.put("/{exercise_id}/weights", response_model=ExerciseWeightsResponse)
def replace_exercise_weights_endpoint(
    exercise_id: int,
    payload: ExerciseWeightsUpdateRequest,
) -> dict[str, Any]:
    try:
        weights = replace_exercise_weights(exercise_id, payload.weights)
    except KeyError as exc:
        raise HTTPException(status_code=404, detail="Exercise not found.") from exc
    except ActiveExerciseWeightError as exc:
        raise HTTPException(status_code=409, detail=str(exc)) from exc
    except ValueError as exc:
        raise HTTPException(status_code=400, detail=str(exc)) from exc

    return {
        "exercise_id": exercise_id,
        "weights": weights,
    }


@profiles_router.get("", response_model=ExerciseProfilesResponse)
def get_exercise_profiles() -> dict[str, list[dict[str, str]]]:
    return {"profiles": list_exercise_profiles()}